fix(ui): split Widgetbook feed stories into catalog_part9 (#2870)
Extract PlayerTurnEventFeedCard Widgetbook stories from
catalog_part7.dart so part fragments stay under the repo.part_unit_size
1000-line ceiling after the mobile-viewport story landed in PR #3204.
