+++ app/lib/widgetbook/catalog_part7.dart
@@ -5,11 +5,12 @@
 //
 // In-game shell chrome stories for issue #2861 S12: top bar, tab bar,
 // bottom-left corner controls, map display options dialog, the player
-// turn-events feed card, the empire left rail (with tooltips), and the
+// empire left rail (with tooltips), and the
 // region minimap (visible / hidden / narrow). Stories for the wide-only
 // players bar, side menu, victory overlay, exit-confirm dialog, game
 // screen, pause menu panel, and the narrow detail overlay slot already
-// live in the earlier catalog parts.
+// live in the earlier catalog parts. Player turn event feed card stories
+// live in `catalog_part9.dart`.
 //
 // Empire left rail and region minimap stories use stand-in providers
 // scoped to this catalog file: a no-op `GameService` that returns a `null`
@@ -228,135 +229,6 @@ List<WidgetbookNode> get gameMapOptionsDialogDirectories => [
             showProvinceOverlay: false,
             showProvinceOwnershipTint: false,
             showProvinceNamesLayer: false,
-          ),
-        ),
-      ),
-    ],
-  ),
-];
-
-/// Player turn event feed card stories.
-/// SPEC/ui/player-turn-event-feed.md. Issue #2861 S7 + S12 story (10)
-/// news feed open / closed. The closed variant just renders the toggle
-/// (already covered by `Game Tab Bar` stories), so this folder focuses on
-/// the floating card surface itself: populated and empty states.
-List<WidgetbookNode> get playerTurnEventFeedCardDirectories => [
-  WidgetbookFolder(
-    name: 'Player Turn Event Feed Card',
-    children: [
-      WidgetbookUseCase(
-        name: 'Populated — three entries (top entry tappable)',
-        builder: (context) => _playerTurnEventFeedCardStoryFrame(
-          child: PlayerTurnEventFeedCard(
-            entries: [
-              PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'Castile completed Castle in Lisbon.',
-                onTap: () {},
-              ),
-              const PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'England declared war on France.',
-              ),
-              const PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'New trade route established: Lisbon ↔ Bordeaux.',
-              ),
-            ],
-            // ignore: avoid_hardcoded_strings_in_widgets
-            emptyLabel: 'No events this turn.',
-          ),
-        ),
-      ),
-      WidgetbookUseCase(
-        name: 'Empty — no events this turn',
-        builder: (context) => _playerTurnEventFeedCardStoryFrame(
-          child: const PlayerTurnEventFeedCard(
-            entries: [],
-            // ignore: avoid_hardcoded_strings_in_widgets
-            emptyLabel: 'No events this turn.',
-          ),
-        ),
-      ),
-      WidgetbookUseCase(
-        name: 'Mobile viewport',
-        builder: (context) => mobileViewport(
-          context,
-          _playerTurnEventFeedCardNarrowStoryFrame(
-            viewportWidth: 360,
-            child: const PlayerTurnEventFeedCard(
-              entries: [
-                PlayerTurnEventFeedEntry(
-                  // ignore: avoid_hardcoded_strings_in_widgets
-                  text: 'Castile completed Castle in Lisbon.',
-                ),
-                PlayerTurnEventFeedEntry(
-                  // ignore: avoid_hardcoded_strings_in_widgets
-                  text: 'England declared war on France.',
-                ),
-              ],
-              // ignore: avoid_hardcoded_strings_in_widgets
-              emptyLabel: 'No events this turn.',
-              narrow: true,
-            ),
-          ),
-        ),
-      ),
-      WidgetbookUseCase(
-        name: 'Narrow (360 dp) — populated, clamp(180, 50vw, 260)',
-        builder: (context) => _playerTurnEventFeedCardNarrowStoryFrame(
-          viewportWidth: 360,
-          child: const PlayerTurnEventFeedCard(
-            entries: [
-              PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'Castile completed Castle in Lisbon.',
-              ),
-              PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'England declared war on France.',
-              ),
-            ],
-            // ignore: avoid_hardcoded_strings_in_widgets
-            emptyLabel: 'No events this turn.',
-            narrow: true,
-          ),
-        ),
-      ),
-      WidgetbookUseCase(
-        name: 'Narrow (460 dp) — populated, 50vw mid-range',
-        builder: (context) => _playerTurnEventFeedCardNarrowStoryFrame(
-          viewportWidth: 460,
-          child: const PlayerTurnEventFeedCard(
-            entries: [
-              PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'Castile completed Castle in Lisbon.',
-              ),
-              PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'England declared war on France.',
-              ),
-              PlayerTurnEventFeedEntry(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                text: 'New trade route established: Lisbon ↔ Bordeaux.',
-              ),
-            ],
-            // ignore: avoid_hardcoded_strings_in_widgets
-            emptyLabel: 'No events this turn.',
-            narrow: true,
-          ),
-        ),
-      ),
-      WidgetbookUseCase(
-        name: 'Narrow (599 dp) — empty, clamp upper bound (260 dp)',
-        builder: (context) => _playerTurnEventFeedCardNarrowStoryFrame(
-          viewportWidth: 599,
-          child: const PlayerTurnEventFeedCard(
-            entries: [],
-            // ignore: avoid_hardcoded_strings_in_widgets
-            emptyLabel: 'No events this turn.',
-            narrow: true,
           ),
         ),
       ),
@@ -606,63 +478,6 @@ class _GameMapOptionsDialogStoryHostState
       ),
     );
   }
-}
-
-/// News feed card frame: float the card against a representative dark map
-/// background scrim and keep the wide-shell width so the chrome reads the
-/// same way it does pinned to the in-game map stack.
-MaterialApp _playerTurnEventFeedCardStoryFrame({required Widget child}) {
-  return MaterialApp(
-    debugShowCheckedModeBanner: false,
-    theme: AppThemes.editorialMonocle,
-    localizationsDelegates: AppLocalizationsBinding.localizationsDelegates,
-    supportedLocales: AppLocalizations.supportedLocales,
-    home: Scaffold(
-      backgroundColor: EditorialMonoclePalette.bgDeep,
-      body: Align(
-        alignment: Alignment.topRight,
-        child: Padding(
-          padding: const EdgeInsets.all(24),
-          child: child,
-        ),
-      ),
-    ),
-  );
-}
-
-/// Narrow-viewport news feed card frame: clamps the visible canvas to a
-/// representative narrow viewport width via `MediaQuery.size`, anchors
-/// the card top-right (matching the production stack placement on
-/// narrow), and forwards the editorial-monocle theme so the chrome
-/// reads identically to the wide story (issue #2870 S3 / Req 11; SPEC
-/// `SPEC/ui/player-turn-event-feed.md` § Card chrome — narrow layout).
-MaterialApp _playerTurnEventFeedCardNarrowStoryFrame({
-  required double viewportWidth,
-  required Widget child,
-}) {
-  return MaterialApp(
-    debugShowCheckedModeBanner: false,
-    theme: AppThemes.editorialMonocle,
-    localizationsDelegates: AppLocalizationsBinding.localizationsDelegates,
-    supportedLocales: AppLocalizations.supportedLocales,
-    home: MediaQuery(
-      data: MediaQueryData(size: Size(viewportWidth, 640)),
-      child: Scaffold(
-        backgroundColor: EditorialMonoclePalette.bgDeep,
-        body: SizedBox(
-          width: viewportWidth,
-          height: 640,
-          child: Align(
-            alignment: Alignment.topRight,
-            child: Padding(
-              padding: const EdgeInsets.all(8),
-              child: child,
-            ),
-          ),
-        ),
-      ),
-    ),
-  );
 }
 
 /// Empire left rail stories. SPEC/ui/empire-buttons.md § Styling (left rail)
